pyproject.toml
[project]
name = "emails"
version = "0.1.0"
description = "Migrate emails from Gmail to Zoho via IMAP"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "python-dotenv>=1.0",
]

[project.scripts]
emails = "emails.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0",
]
